pyproject.toml
[project]
name = "virality-coach"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chardet>=5.2.0",
    "charset-normalizer>=3.4.2",
    "ffmpeg-python>=0.2.0",
    "google-generativeai>=0.8.5",
    "openai>=1.95.1",
    "openai-whisper>=20250625",
    "pillow>=11.3.0",
    "plotly>=6.2.0",
    "python-dotenv>=1.1.1",
    "scenedetect[opencv]>=0.6.2",
    "sievedata>=1.5.0",
    "streamlit>=1.46.1",
    "yt-dlp>=2025.7.21",
]
